+++ src/screens/worker/mypage/ProfileEditScreen.tsx
@@ -6,7 +6,7 @@ import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import HomeBackButton from "../../../components/common/HomeBackButton";
 import { Text } from "../../../components/common/Text";
 import Header from "../../../components/layout/Header";
-import MyPageDrawer from "../../../components/worker/mypage/MyPageDrawer";
+import MyPageDrawer from "../../../components/mypage/drawer/MyPageDrawer";
 import { WorkerStackParamList } from "../../../navigation/WorkerStack";
 import ProfileFieldRow from "../../../components/mypage/profileEdit/ProfileFieldRow";
 import ProfilePhoto from "../../../components/mypage/profileEdit/ProfilePhoto";
