+++ src/screens/worker/mypage/WithdrawScreen.tsx
@@ -6,7 +6,7 @@ import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import HomeBackButton from "../../../components/common/HomeBackButton";
 import { Text } from "../../../components/common/Text";
 import Header from "../../../components/layout/Header";
-import MyPageDrawer from "../../../components/worker/mypage/MyPageDrawer";
+import MyPageDrawer from "../../../components/mypage/drawer/MyPageDrawer";
 import { WorkerStackParamList } from "../../../navigation/WorkerStack";
 
 type Props = NativeStackScreenProps<WorkerStackParamList, "Withdraw">;
